+++ src/components/mod.rs
@@ -1,9 +1,11 @@
+mod chips;
 mod icon;
 mod state_layer;
 mod surface;
 mod typography;
 
 pub use self::{
+    chips::{AssistChip, FilterChip, InputChip},
     icon::{Icon, IconData, IconStyle},
     state_layer::StateLayer,
     surface::Surface,